(Re)Move fmt
