+++ include/zq/config.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <fmt/format.h>
+
 #include <zmq.h>
 #include <tl/expected.hpp>
